+++ packages/lib/uni-data-checkbox/uni-data-checkbox.vue
@@ -65,15 +65,9 @@
 	 * @event {Function} change  选中发生变化触发
 	 */
 
-	// import clientdb from './clientdb.js'
 	export default {
 		name: 'uniDataChecklist',
-		// mixins: [clientdb],
 		mixins: [uniCloud.mixinDatacom || {}],
-		// model: {
-		// 	prop: 'modelValue',
-		// 	event: 'update:modelValue'
-		// },
 		emits:['input','update:modelValue','change'],
 		props: {
 			mode: {
@@ -160,11 +154,18 @@
 			},
 			value(newVal) {
 				this.dataList = this.getDataList(newVal)
-				this.formItem && this.formItem.setValue(newVal)
+				// fix by mehaotian is_reset 在 uni-forms 中定义
+				if(!this.is_reset){
+					this.is_reset = false
+					this.formItem && this.formItem.setValue(newVal)
+				}
 			},
 			modelValue(newVal) {
 				this.dataList = this.getDataList(newVal);
-				this.formItem && this.formItem.setValue(newVal);
+				if(!this.is_reset){
+					this.is_reset = false
+					this.formItem && this.formItem.setValue(newVal)
+				}
 			}
 		},
 		data() {
@@ -178,8 +179,8 @@
 				},
 				isLocal:true,
 				styles: {
-					selectedColor: '#007aff',
-					selectedTextColor: '#333',
+					selectedColor: '#2979ff',
+					selectedTextColor: '#666',
 				},
 				isTop:0
 			};
@@ -199,6 +200,11 @@
 			if (this.formItem) {
 				this.isTop = 6
 				if (this.formItem.name) {
+					// 如果存在name添加默认值,否则formData 中不存在这个字段不校验
+					if(!this.is_reset){
+						this.is_reset = false
+						this.formItem.setValue(this.dataValue)
+					}
 					this.rename = this.formItem.name
 					this.form.inputChildrens.push(this)
 				}
@@ -377,7 +383,7 @@
 			 */
 			setStyleBackgroud(item) {
 				let styles = {}
-				let selectedColor = this.selectedColor?this.selectedColor:'#007aff'
+				let selectedColor = this.selectedColor?this.selectedColor:'#2979ff'
 				if (this.mode !== 'list') {
 					styles['border-color'] = item.selected?selectedColor:'#DCDFE6'
 				}
@@ -393,7 +399,7 @@
 			setStyleIcon(item) {
 				let styles = {}
 				let classles = ''
-				let selectedColor = this.selectedColor?this.selectedColor:'#007aff'
+				let selectedColor = this.selectedColor?this.selectedColor:'#2979ff'
 				styles['background-color'] = item.selected?selectedColor:'#fff'
 				styles['border-color'] = item.selected?selectedColor:'#DCDFE6'
 
@@ -410,11 +416,11 @@
 			setStyleIconText(item) {
 				let styles = {}
 				let classles = ''
-				let selectedColor = this.selectedColor?this.selectedColor:'#007aff'
+				let selectedColor = this.selectedColor?this.selectedColor:'#2979ff'
 				if (this.mode === 'tag') {
-					styles.color = item.selected?(this.selectedTextColor?this.selectedTextColor:'#fff'):'#333'
+					styles.color = item.selected?(this.selectedTextColor?this.selectedTextColor:'#fff'):'#666'
 				} else {
-					styles.color = item.selected?(this.selectedTextColor?this.selectedTextColor:selectedColor):'#333'
+					styles.color = item.selected?(this.selectedTextColor?this.selectedTextColor:selectedColor):'#666'
 				}
 				if(!item.selected && item.disabled){
 					styles.color = '#999'
@@ -442,7 +448,7 @@
 </script>
 
 <style lang="scss">
-	$checked-color: #007aff;
+	$checked-color: #2979ff;
 	$border-color: #DCDFE6;
 	$disable:0.4;
 
@@ -465,7 +471,7 @@
 	.uni-data-checklist {
 		position: relative;
 		z-index: 0;
-
+		flex: 1;
 		// 多选样式
 		.checklist-group {
 			@include flex;
@@ -498,7 +504,7 @@
 					justify-content: space-between;
 					.checklist-text {
 						font-size: 14px;
-						color: #333;
+						color: #666;
 						margin-left: 5px;
 						line-height: 14px;
 					}
@@ -529,7 +535,7 @@
 					width: 16px;
 					height: 16px;
 					border: 1px solid $border-color;
-					border-radius: 2px;
+					border-radius: 4px;
 					background-color: #fff;
 					z-index: 1;
 					.checkbox__inner-icon {
@@ -645,7 +651,7 @@
 				// 按钮样式
 				&.is--button {
 					margin-right: 10px;
-					padding: 5px 15px;
+					padding: 5px 10px;
 					border: 1px $border-color solid;
 					border-radius: 3px;
 					transition: border-color 0.2s;
@@ -717,7 +723,7 @@
 
 					.checklist-text {
 						margin: 0;
-						color: #333;
+						color: #666;
 					}
 
 					// 禁用
@@ -777,7 +783,11 @@
 								transform: rotate(45deg);
 							}
 						}
-
+						.radio__inner {
+							.radio__inner-icon {
+								opacity: 1;
+							}
+						}
 						.checklist-text {
 							color: $checked-color;
 						}