+++ packages/lib/uni-search-bar/uni-search-bar.vue
@@ -1,30 +1,35 @@
-<template>
+s<template>
 	<view class="uni-searchbar">
 		<view :style="{borderRadius:radius+'px',backgroundColor: bgColor}" class="uni-searchbar__box" @click="searchClick">
 			<view class="uni-searchbar__box-icon-search">
 				<slot name="searchIcon">
-					<uni-icons color="#999999" size="18" type="search" />
+					<uni-icons color="#c0c4cc" size="18" type="search" />
 				</slot>
 			</view>
-			<input v-if="show || searchVal" :focus="showSync" :placeholder="placeholder" :maxlength="maxlength" class="uni-searchbar__box-search-input"
+			<input v-if="show || searchVal" :focus="showSync" :placeholder="placeholderText" :maxlength="maxlength" class="uni-searchbar__box-search-input"
 			 confirm-type="search" type="text" v-model="searchVal" @confirm="confirm" @blur="blur" @focus="emitFocus" />
 			<text v-else class="uni-searchbar__text-placeholder">{{ placeholder }}</text>
 			<view v-if="show && (clearButton==='always'||clearButton==='auto'&&searchVal!=='')" class="uni-searchbar__box-icon-clear"
 			 @click="clear">
 				<slot name="clearIcon">
-					<uni-icons color="#c0c4cc" size="15" type="clear" />
+					<uni-icons color="#c0c4cc" size="20" type="clear" />
 				</slot>
 			</view>
 		</view>
-		<text @click="cancel" class="uni-searchbar__cancel" v-if="cancelButton ==='always' || show && cancelButton ==='auto'">{{cancelText}}</text>
+		<text @click="cancel" class="uni-searchbar__cancel" v-if="cancelButton ==='always' || show && cancelButton ==='auto'">{{cancelTextI18n}}</text>
 	</view>
 </template>
 
 <script>
+	import {
+	initVueI18n
+	} from '@dcloudio/uni-i18n'
+	import messages from './i18n/index.js'
+	const {	t	} = initVueI18n(messages)
 
 	/**
 	 * SearchBar 搜索栏
-	 * @description 评分组件
+	 * @description 搜索栏组件，通常用于搜索商品、文章等
 	 * @tutorial https://ext.dcloud.net.cn/plugin?id=866
 	 * @property {Number} radius 搜索栏圆角
 	 * @property {Number} maxlength 输入最大长度
@@ -53,7 +58,7 @@
 		props: {
 			placeholder: {
 				type: String,
-				default: "请输入搜索内容"
+				default: ""
 			},
 			radius: {
 				type: [Number, String],
@@ -99,7 +104,16 @@
 				searchVal: ''
 			}
 		},
+		computed:{
+			cancelTextI18n() {
+				return this.cancelText || t("uni-search-bar.cancel")
+			},
+			placeholderText() {
+				return this.placeholder || t("uni-search-bar.placeholder")
+			}
+		},
 		watch: {
+			// #ifndef VUE3
 			value: {
 				immediate: true,
 				handler(newVal) {
@@ -109,6 +123,8 @@
 					}
 				}
 			},
+			// #endif
+			// #ifdef VUE3
 			modelValue: {
 				immediate: true,
 				handler(newVal) {
@@ -118,6 +134,7 @@
 					}
 				}
 			},
+			// #endif
 			focus: {
 				immediate: true,
 				handler(newVal) {
@@ -130,8 +147,12 @@
 				}
 			},
 			searchVal(newVal, oldVal) {
+				// #ifndef VUE3
 				this.$emit("input", newVal)
+				// #endif
+				// #ifdef VUE3
 				this.$emit("update:modelValue", newVal)
+				// #endif
 			}
 		},
 		methods: {
@@ -202,8 +223,8 @@
 		/* #endif */
 		flex-direction: row;
 		position: relative;
-		padding: $uni-spacing-col-base;
-		// background-color: $uni-bg-color;
+		padding: 10px;
+		// background-color: #fff;
 	}
 
 	.uni-searchbar__box {
@@ -219,9 +240,6 @@
 		align-items: center;
 		height: $uni-searchbar-height;
 		padding: 5px 8px 5px 0px;
-		border-width: 0.5px;
-		border-style: solid;
-		border-color: $uni-border-color;
 	}
 
 	.uni-searchbar__box-icon-search {
@@ -233,13 +251,13 @@
 		padding: 0 8px;
 		justify-content: center;
 		align-items: center;
-		color: $uni-text-color-placeholder;
+		color: #B3B3B3;
 	}
 
 	.uni-searchbar__box-search-input {
 		flex: 1;
-		font-size: $uni-font-size-base;
-		color: $uni-text-color;
+		font-size: 14px;
+		color: #333;
 	}
 
 	.uni-searchbar__box-icon-clear {
@@ -252,8 +270,8 @@
 	}
 
 	.uni-searchbar__text-placeholder {
-		font-size: $uni-font-size-base;
-		color: $uni-text-color-placeholder;
+		font-size: 14px;
+		color: #B3B3B3;
 		margin-left: 5px;
 	}
 
@@ -261,7 +279,7 @@
 		padding-left: 10px;
 		line-height: $uni-searchbar-height;
 		font-size: 14px;
-		color: $uni-text-color;
+		color: #333333;
 		/* #ifdef H5 */
 		cursor: pointer;
 		/* #endif */