+++ uni_modules/uni-ui/changelog.md
@@ -1,3 +1,70 @@
+## 1.4.7（2021-11-19）
+- uni-ui 新增 支持国际化
+- uni-ui 优化 组件UI，并提供设计资源，详见:[https://uniapp.dcloud.io/component/uniui/resource](https://uniapp.dcloud.io/component/uniui/resource)
+- uni-ui 文档迁移，详见:[https://uniapp.dcloud.io/component/uniui/uni-ui](https://uniapp.dcloud.io/component/uniui/uni-ui)
+- uni-badge 修改 size 属性默认值调整为 small
+- uni-badge 修改 type 属性，默认值调整为 error，info 替换 default
+- uni-badge 修复 在字节小程序上样式不生效的 bug
+- uni-calendar 修复 弹出层被 tabbar 遮盖 bug
+- uni-card 重构插槽的用法 ，header 替换为 title 
+- uni-card 新增 actions 插槽
+- uni-card 新增 cover 封面图属性和插槽
+- uni-card 新增 padding 内容默认内边距离
+- uni-card 新增 margin 卡片默认外边距离
+- uni-card 新增 spacing 卡片默认内边距
+- uni-card 新增 shadow 卡片阴影属性
+- uni-card 取消 mode 属性，可使用组合插槽代替
+- uni-card 取消 note 属性 ，使用actions插槽代替
+- uni-collapse 优化 show-arrow 属性默认为true
+- uni-collapse 新增 show-arrow 属性，控制是否显示右侧箭头
+- uni-countdown 新增 font-size 支持自定义字体大小
+- uni-data-checkbox 修复 在uni-forms中 modelValue 中不存在当前字段，当前字段必填写也不参与校验的问题
+- uni-data-checkbox 修复 单选 list 模式下 ，icon 为 left 时，选中图标不显示的问题
+- uni-data-checkbox 修复 在 uni-forms 中重置表单，错误信息无法清除的问题
+- uni-dateformat 优化 默认时间不再是当前时间，而是显示'-'字符
+- uni-datetime-picker 修复 hide-second 在移动端的 bug
+- uni-datetime-picker 修复 单选赋默认值时，赋值日期未高亮的 bug
+- uni-datetime-picker 修复 赋默认值时，移动端未正确显示时间的 bug
+- uni-datetime-picker 新增 hide-second 属性，支持只使用时分，隐藏秒
+- uni-datetime-picker 优化 取消选中时（范围选）直接开始下一次选择, 避免多点一次
+- uni-datetime-picker 优化 移动端支持清除按钮，同时支持通过 ref 调用组件的 clear 方法
+- uni-datetime-picker 优化 调整字号大小，美化日历界面
+- uni-datetime-picker 优化 范围选择器在 pc 端过宽的问题
+- uni-datetime-picker 新增 支持作为 uni-forms 子组件相关功能
+- uni-datetime-picker 修复 在 uni-forms 中使用时，选择时间报 NAN 错误的 bug
+- uni-datetime-picker 修复 type 属性动态赋值无效的 bug
+- uni-datetime-picker 修复 ‘确认’按钮被 tabbar 遮盖 bug
+- uni-datetime-picker 修复 组件未赋值时范围选左、右日历相同的 bug
+- uni-datetime-picker 修复 范围选未正确显示当前值的 bug
+- uni-datetime-picker 修复 h5 平台（移动端）报错 'cale' of undefined 的 bug
+- uni-easyinput 修复 在 uni-forms 的动态表单中默认值校验不通过的 bug
+- uni-easyinput 修复 在 uni-forms 中重置表单，错误信息无法清除的问题
+- uni-file-picker 新增 参数中返回 fileID 字段
+- uni-file-picker 修复 腾讯云传入fileID 不能回显的bug
+- uni-file-picker 修复 选择图片后，不能放大的问题
+- uni-file-picker 修复 由于 0.2.11 版本引起的不能回显图片的Bug
+- uni-file-picker 新增 clearFiles(index) 方法，可以手动删除指定文件
+- uni-file-picker 修复 v-model 值设为 null 报错的Bug
+- uni-file-picker 修复 return-type="object" 时，无法删除文件的Bug
+- uni-file-picker 修复 auto-upload 属性失效的Bug
+- uni-forms 修复 label 插槽不生效的bug
+- uni-forms 修复 没有添加校验规则的字段依然报错的Bug
+- uni-forms 修复 重置表单错误信息无法清除的问题
+- uni-forms 修复 表单验证只生效一次的问题
+- uni-icons 新增 更多图标
+- uni-icons 优化 自定义图标使用方式
+- uni-link 修复 在 nvue 下不显示的 bug
+- uni-pagination 修复 current 、value 属性未监听，导致高亮样式失效的 bug
+- uni-rate 优化 默认值修改为 0 颗星
+- uni-search-bar 修复 value 属性与 modelValue 属性不兼容的Bug
+- uni-swipe-action 新增 close-all 方法，关闭所有已打开的组件
+- uni-swipe-action 新增 resize() 方法，在非微信小程序、h5、app-vue端出现不能滑动的问题的时候，重置组件
+- uni-swipe-action 修复 app 端偶尔出现类似 Page[x][-x,xx;-x,xx,x,x-x] 的问题 
+- uni-swipe-action 优化 微信小程序、h5、app-vue 滑动逻辑，避免出现动态新增组件后不能滑动的问题
+- uni-tag 新增 提供组件设计资源，组件样式调整
+- uni-tag 移除 插槽
+- uni-tag 移除 type 属性的 royal 选项
+- uni-tag type 不是 default 时，size 为 small 字体大小显示不正确
 ## 1.4.2（2021-08-20）
 - 新增 uni-ui 组件支持国际化 i18n
 - uni-collapse 优化 show-arrow 属性默认为true